+++ CHANGELOG.md
@@ -18,7 +18,7 @@ Some guidelines in reading this document:
   * **[BC]** - Ommits `localResources`, now that a denormalization methods exists.
   * Exposes `selectRequestRaw`, so a request might be selected without denormalization step
   * Fix [#12](https://github.com/log-oscon/redux-wpapi/issues/12) – Promise returned from middleware dispatch was overlaping responses with same name
-  * Fix [#14](https://github.com/log-oscon/redux-wpapi/issues/14) - lastUpdatedCache should be used as Symbol for preventing collisions, now been tested
+  * Fix [#14](https://github.com/log-oscon/redux-wpapi/issues/14) - lastCacheUpdate should be used as Symbol for preventing collisions, now been tested
 
 ## 1.1.0